Initial commit with basic movement of the players
Board is ready and so far displayed correctly. Board has a Sorted Dictionary of Positions containing data for each field. Game is created for a Player and CPU to check the status of the features. Player can move normally, but more of the game rules how to be implemenmted. The same applies to CPU. Next step is to work on capturing of the pieces