Memento, Undo-Redo, Kings, Winning the game
The Memento pattern was used to untile undo-redo options. Memento pattern allows to store states of the game board with all the positions of the pieces. List of all states is kept in GameHistory class, which serves as name indicates and will be used to implement Replay Game feature
The kings move and capture correctly. Checking a winner is working and game exits correctly.