Multiple jumps implemented
Finding legal moves method was split into smaller parts. Checking for capturing was moved to be checked first as this allows easy implementation of multiple jumps.
Capturing piece variable was added to mark the piece that had capturing move as this is used later to check for more capturing moves for this piece only